HttpRequest 책임 분리
HttpRequestLine 클래스를 만들어서 requestLine과 관련된 데이터만 관리하도록 함
HttpRequestHeader 클래스를 만들어서 RequestHeader의 데이터만 관리하도록 함
